+++ vpype/config.py
@@ -42,13 +42,18 @@ class PaperConfig:
     """Data class containing configuration for a give plotter type/paper size combinations."""
 
     name: str  #: name of the paper format
-    paper_size: Tuple[float, float]  #: X/Y axis convention of the plotter
-    x_range: Tuple[int, int]  #: admissible range of X coordinates
-    y_range: Tuple[int, int]  #: admissible range of Y coordinates
     y_axis_up: bool  #: if True, the Y axis point upwards instead of downwards
     origin_location: Tuple[
         float, float
     ]  #: location on paper of the (0, 0) plotter unit coordinates
+
+    paper_size: Optional[Tuple[float, float]] = None  #: X/Y axis convention of the plotter
+    paper_orientation: Optional[
+        str
+    ] = None  #: orientation of the plotter coordinate system on paper
+    x_range: Optional[Tuple[int, int]] = None  #: admissible range of X coordinates
+    y_range: Optional[Tuple[int, int]] = None  #: admissible range of Y coordinates
+    origin_location_reference: Optional[str] = "topleft"  #: reference for ``origin_location``
 
     info: str = ""  #: information printed to the user when paper is used
     rotate_180: bool = False  #: if True, the geometries are rotated by 180 degrees on the page
@@ -64,11 +69,15 @@ class PaperConfig:
     def from_config(cls, data: Dict[str, Any]) -> "PaperConfig":
         return cls(
             name=data["name"],
-            paper_size=_convert_length_pair(data["paper_size"]),
-            x_range=(data["x_range"][0], data["x_range"][1]),
-            y_range=(data["y_range"][0], data["y_range"][1]),
             y_axis_up=data["y_axis_up"],
             origin_location=_convert_length_pair(data["origin_location"]),
+            paper_size=_convert_length_pair(data["paper_size"])
+            if "paper_size" in data
+            else None,
+            paper_orientation=data.get("paper_orientation", None),
+            x_range=(data["x_range"][0], data["x_range"][1]) if "x_range" in data else None,
+            y_range=(data["y_range"][0], data["y_range"][1]) if "y_range" in data else None,
+            origin_location_reference=data.get("origin_location_reference", "topleft"),
             info=data.get("info", ""),
             rotate_180=data.get("rotate_180", False),
             set_ps=data.get("set_ps", None),
@@ -83,7 +92,7 @@ class PlotterConfig:
 
     name: str  #: name of the plotter
     paper_configs: List[PaperConfig]  #: list of :class:`PaperConfig` instance
-    plotter_unit_length: float  #: phyiscal size of plotter units (in pixel)
+    plotter_unit_length: float  #: physical size of plotter units (in pixel)
     pen_count: int  #: number of pen supported by the plotter
 
     info: str = ""  #: information printed to the user when plotter is used