+++ vpype_viewer/shaders/fast_line_fragment.glsl
@@ -1,0 +1,9 @@
+#version 330
+
+flat in vec4 v_color;
+
+out vec4 f_color;
+
+void main() {
+    f_color = v_color;
+}